+++ python/tests/test_summary.py
@@ -66,6 +66,7 @@ class TestSummary(unittest.TestCase):
         with io.BytesIO() as out:
             s.write(out, format="yaml", annotate=True)
             self.assertEqual(out.getvalue()[:11], b"SummaryItem")
+            self.assertIn(b'# sfc Surface (of the Earth, which includes sea surface)', out.getvalue())
 
             out.seek(0)
             s1 = arki.Summary.read_yaml(out)
@@ -77,17 +78,32 @@ class TestSummary(unittest.TestCase):
             s.write(out, format="json")
             self.assertEqual(out.getvalue()[:21], b'{"items":[{"origin":{')
 
+            # Read from bytes()
+            s1 = arki.Summary.read_json(out.getvalue())
+            self.assertCountEqual(s.to_python()["items"], s1.to_python()["items"])
+
+            # Read from str()
+            s1 = arki.Summary.read_json(out.getvalue().decode())
+            self.assertCountEqual(s.to_python()["items"], s1.to_python()["items"])
+
+            # Read from binary abstract FD
             out.seek(0)
             s1 = arki.Summary.read_json(out)
-            self.assertEqual(s.to_python(), s1.to_python())
+            self.assertCountEqual(s.to_python()["items"], s1.to_python()["items"])
+
+            # Read from string abstract FD
+            with io.StringIO(out.getvalue().decode()) as infd:
+                s1 = arki.Summary.read_json(infd)
+                self.assertCountEqual(s.to_python()["items"], s1.to_python()["items"])
 
         with io.BytesIO() as out:
             s.write(out, format="json", annotate=True)
             self.assertEqual(out.getvalue()[:21], b'{"items":[{"origin":{')
+            self.assertIn(b'"desc":"', out.getvalue())
 
             out.seek(0)
             s1 = arki.Summary.read_json(out)
-            self.assertEqual(s.to_python(), s1.to_python())
+            self.assertCountEqual(s.to_python()["items"], s1.to_python()["items"])
 
     def test_write_short(self):
         s = self.read("inbound/test.grib1")