+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added `format_metadata.rst` to Python HOWTOs (#230)
 * Python: added `arkimet.Summary.read_binary` (#230)
 * Python: added `arkimet.Summary.read_yaml` (#230)
+* Python: added `arkimet.Summary.read_json` (#230)
 * Python: added `annotate: bool = False` to `arkimet.Summary.write` (#230)
 
 # New in version 1.27